push all the updates into repo
