+++ app/README.md
@@ -1,0 +1,6 @@
+---
+title: Nano-AutoGrad
+app_file: app.py
+sdk: gradio
+sdk_version: 3.35.2
+---
