+++ app/init.py
@@ -1,0 +1,5 @@
+import os 
+import sys
+# Add the parent directory of app.py to the system path
+parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+sys.path.append(parent_dir)
