+++ app/requirements.txt
@@ -1,0 +1,16 @@
+pandas
+plotly
+numpy 
+matplotlib
+random2
+scipy 
+scikit-learn
+kaleido
+ipython
+nbformat>=4.2.0
+ipywidgets>=7.0.0
+imageio
+graphviz
+gradio
+nano-autograds
+sphinx
